added database and hashing, including hash checking
Added flask-login to the code, including @login.user_loader decorator, and replaced the fake login (flash) with
is_authenticated: a property that is True if the user has valid credentials or False otherwise.
is_active: a property that is True if the user's account is active or False otherwise.
is_anonymous: a property that is False for regular users, and True for a special, anonymous user.
get_id(): a method that returns a unique identifier for the user as a string (unicode, if using Python 2).

added logout to routes
and to base.html

require login prior to seeing protected pages, initially just / and /index.

Handling the next for redirect backs.

Added registration
- to forms as a class
- added registration.html template

added link to registration page from login page (login.html template)

Signed-off-by: robertkirkpatrick <rk3505@nyu.edu>